liri-python-app-master/requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
blessings==1.7
certifi==2019.3.9
chardet==3.0.4
idna==2.5
idna-ssl==1.1.0
inquirer==2.5.1
multidict==4.5.2
readchar==2.0.1
requests2==2.16.0
six==1.12.0
spotify==0.3.0
twitter==1.18.0
typing-extensions==3.7.2
urllib3==1.21.1
yarl==1.3.0
